[Fix] get raw href from the sieve and not a strangely encoded one.
The sieve could fail marking an item as read because the js code was
receiving a url that didn't match the one registered in the app (and
initially sent with the document).

Apparently (on firefox at least) calling a DOM's element.href applies
some encoding to the actual string value of href present in the HTML
code such that we would end-up marking the following as read:
http://scripting.com/2020/01/07/155714.html?title=Sports%20is%20judgmental%2C%20and%20that%27s%20good."

whereas the actual href value in the html code (and expected by the
server) was:
http://scripting.com/2020/01/07/155714.html?title=Sports%20is%20judgmental%2C%20and%20that's%20good."

Calling .getAttributes("href") works around this automatic encoding it
seems and allows getting the original value.