Fix collation for feed with broken dates.
Items of a feed were dates are broken typically end up with all items
at the same date and if on top of that there are too many items (yes
it happened), then the collation is taking over and over the same
items when it as supposed to take new chunks of the items list at each
iteration. The caveat now is that will ignore items in this kind of
situation but those items were never caught anyway with the bug.